Crazy update.
Fixed the bullet velocity.
Made the jframe bigger and changed the menus and map to match.
updated the character movement to a more streamlined way.
Was able to get the game timer to display in the console the final game time.
Made a pause menu that can take you back to the main menu or to the game.( doesnt actually pause the game) (cant figure out how to pause the game timer thread.)
Updated the camera movement to match the new screen size.